Add lazy evaluation and dynamic node expansion
This PR inverts the execution model -- from recursively calling nodes to
using a topological sort of the nodes. This change allows for
modification of the node graph during execution. This allows for two
major advantages:
1. The implementation of lazy evaluation in nodes. For example, if a
   "Mix Images" node has a mix factor of exactly 0.0, the second image
   input doesn't even need to be evaluated (and visa-versa if the mix
   factor is 1.0).
2. Dynamic expansion of nodes. This allows for the creation of dynamic
   "node groups". For example, I implemented an "Advanced Prompt" node.
   It takes a string prompt, but will also dynamically create LoraLoader
   nodes for any loras mentioned in the prompt.

Note that this branch is currently a work-in-progress and probably isn't
suitable for use yet.